+++ nova2d/nova-zero/src/core/Game.h
@@ -83,7 +83,7 @@ namespace novazero
 			static void ConfigureDebugPosition(Vec2Int pos) { if (s_DebugOverlay) { s_DebugOverlay->SetPositionInt(pos); } }
 
 			static double s_DeltaTime;
-			static double GetDeltaTime() { return s_DeltaTime; }
+			static float GetDeltaTime() { return (float)s_DeltaTime; }
 			static double s_FPS;
 
 			static int s_Width;
