+++ nova2d/nova-zero/src/graphics/Sprite.cpp
@@ -194,8 +194,8 @@ namespace novazero
 			if (GetDrawable())
 				GetDrawable()->SetDrawScale(scale);
 
-			m_DestRect.w = m_FrameSize.x * scale;
-			m_DestRect.h = m_FrameSize.y * scale;
+			m_DestRect.w = (int)(m_FrameSize.x * scale);
+			m_DestRect.h = (int)(m_FrameSize.y * scale);
 		}
 
 		void Sprite::Draw(float oX, float oY)
